misc(ultima): code review updates

+++ db/migrate/20260915080340_add_ultima_application_id_to_request_metadata.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 class AddUltimaApplicationIdToRequestMetadata < ActiveRecord::Migration[8.1]
   def change
-    add_column :request_metadata, :ultima_application_id, :integer
+    add_column :request_metadata, :ultima_application_id, :bigint
   end
 end
